+++ torchtext/experimental/transforms.py
@@ -9,7 +9,6 @@ from torchtext._torchtext import RegexTokenizer as RegexTokenizerPybind, Sentenc
 
 __all__ = [
     "basic_english_normalize",
-    "regex_tokenizer",
     "BasicEnglishNormalize",
     "PRETRAINED_SP_MODEL",
     "load_sp_model",
@@ -69,30 +68,6 @@ def basic_english_normalize():
     patterns = [pair[0] for pair in patterns_list]
     replacements = [pair[1] for pair in patterns_list]
     return BasicEnglishNormalize(RegexTokenizerPybind(patterns, replacements, True))
-
-
-def regex_tokenizer(patterns_list):
-    r"""Regex tokenizer for a string sentence that applies all regex replacements defined in patterns_list.
-
-    Args:
-        patterns_list (List[Tuple[str, str]]): a list of tuples (ordered pairs) which contain the regex pattern string
-        as the first element and the replacement string as the second element.
-
-    Examples:
-        >>> import torch
-        >>> from torchtext.experimental.transforms import regex_tokenizer
-        >>> test_sample = 'Basic Regex Tokenization for a Line of Text'
-        >>> patterns_list = [
-            (r'\'', ' \'  '),
-            (r'\"', '')]
-        >>> reg_tokenizer = regex_tokenizer(patterns_list)
-        >>> jit_reg_tokenizer = torch.jit.script(reg_tokenizer)
-        >>> tokens = jit_reg_tokenizer(test_sample)
-    """
-
-    patterns = [pair[0] for pair in patterns_list]
-    replacements = [pair[1] for pair in patterns_list]
-    return RegexTokenizer(RegexTokenizerPybind(patterns, replacements, False))
 
 
 class BasicEnglishNormalize(nn.Module):
